Update 16/11/19 - 22:21
- New collision and moving system
- Add monsters

+++ Cube 2/src/main/java/com/torpill/game/component/Window.java
@@ -164,7 +164,7 @@ public class Window extends JFrame implements Runnable, ComponentListener, Windo
 		return this.height;
 	}
 
-	public void setOffset(int x, int y) {
+	public void setOffset(double x, double y) {
 
 		this.playpan.setOffset(x, y);
 	}
